move text-based adapter to the web socket adapter interface file
