+++ src/index.ts
@@ -29,4 +29,3 @@ export * from "./client";
 export * from "./session";
 export * from "./web_socket_adapter";
 export * from "./web_socket_adapter_pb";
-export * from "./web_socket_adapter_text";
